backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.2
psycopg2-binary==2.9.9

# Auth / Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.4.0
pydantic[email]==2.9.1

# AI / ML
scikit-learn==1.5.2
xgboost==2.1.1
numpy==1.26.4
pandas==2.2.2

# LangChain / LangGraph
langchain==0.3.1
langchain-core==0.3.6
langgraph==0.2.27
langchain-google-genai==2.0.0
langchain-openai==0.2.1

# Utilities
python-dateutil==2.9.0
httpx==0.27.2
redis==5.1.1

# Dev / Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
